requirements.txt
Flask==2.1.0
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
PyMySQL==1.0.2
python-dotenv==0.19.2
Werkzeug==2.1.0
PyJWT==2.3.0
SQLAlchemy==1.4.47


